+++ packages/@uppy/aws-s3-multipart/src/index.js
@@ -241,7 +241,7 @@ class HTTPCommunicationQueue {
 
   async #nonMultipartUpload (file, chunk, signal) {
     const {
-      method = 'post',
+      method = 'POST',
       url,
       fields,
       headers,
@@ -249,7 +249,7 @@ class HTTPCommunicationQueue {
 
     let body
     const data = chunk.getData()
-    if (method === 'post') {
+    if (method.toUpperCase() === 'POST') {
       const formData = new FormData()
       Object.entries(fields).forEach(([key, value]) => formData.set(key, value))
       formData.set('file', data)
@@ -662,7 +662,13 @@ export default class AwsS3Multipart extends BasePlugin {
 
         // NOTE This must be allowed by CORS.
         const etag = ev.target.getResponseHeader('ETag')
+        const location = ev.target.getResponseHeader('Location')
 
+        if (method.toUpperCase() === 'POST' && location === null) {
+          // Not being able to read the Location header is not a fatal error.
+          // eslint-disable-next-line no-console
+          console.warn('AwsS3/Multipart: Could not read the Location header. This likely means CORS is not configured correctly on the S3 Bucket. See https://uppy.io/docs/aws-s3-multipart#S3-Bucket-Configuration for instructions.')
+        }
         if (etag === null) {
           reject(new Error('AwsS3/Multipart: Could not read the ETag header. This likely means CORS is not configured correctly on the S3 Bucket. See https://uppy.io/docs/aws-s3-multipart#S3-Bucket-Configuration for instructions.'))
           return
@@ -671,6 +677,7 @@ export default class AwsS3Multipart extends BasePlugin {
         onComplete?.(etag)
         resolve({
           ETag: etag,
+          ...(location ? { location } : undefined),
         })
       })