+++ packages/@uppy/companion/src/server/emitter/index.js
@@ -8,9 +8,9 @@ let emitter
  * Used to transmit events (such as progress, upload completion) from controllers,
  * such as the Google Drive 'get' controller, along to the client.
  */
-module.exports = (redisUrl, redisPubSubScope) => {
+module.exports = (redisClient, redisPubSubScope) => {
   if (!emitter) {
-    emitter = redisUrl ? redisEmitter(redisUrl, redisPubSubScope) : nodeEmitter()
+    emitter = redisClient ? redisEmitter(redisClient, redisPubSubScope) : nodeEmitter()
   }
 
   return emitter