+++ packages/@uppy/companion/src/standalone/helper.js
@@ -146,9 +146,18 @@ const getConfigFromEnv = () => {
     filePath: process.env.COMPANION_DATADIR,
     redisUrl: process.env.COMPANION_REDIS_URL,
     redisPubSubScope: process.env.COMPANION_REDIS_PUBSUB_SCOPE,
-    // adding redisOptions to keep all companion options easily visible
     //  redisOptions refers to https://www.npmjs.com/package/redis#options-object-properties
-    redisOptions: {},
+    redisOptions: (() => {
+      try {
+        if (!process.env.COMPANION_REDIS_OPTIONS) {
+          return undefined
+        }
+        return JSON.parse(process.env.COMPANION_REDIS_OPTIONS)
+      } catch (e) {
+        logger.warn('COMPANION_REDIS_OPTIONS parse error', e)
+      }
+      return undefined
+    })(),
     sendSelfEndpoint: process.env.COMPANION_SELF_ENDPOINT,
     uploadUrls: uploadUrls ? uploadUrls.split(',') : null,
     secret: getSecret('COMPANION_SECRET'),