+++ packages/@uppy/progress-bar/types/index.d.ts
@@ -1,6 +1,6 @@
-import type { PluginOptions, UIPlugin, PluginTarget } from '@uppy/core'
+import type { PluginTarget, UIPlugin, UIPluginOptions } from '@uppy/core'
 
-export interface ProgressBarOptions extends PluginOptions {
+export interface ProgressBarOptions extends UIPluginOptions {
   target?: PluginTarget
   hideAfterFinish?: boolean
   fixed?: boolean